Add unit test for geometry import
